@@ -4,10 +4,10 @@
   "description": "The boilerplate of react with typescript.",
   "main": "index.js",
   "scripts": {
-    "build": "rimraf dist && cross-env NODE_ENV=production webpack",
-    "format": "prettier --write \"src/**/*.{ts,tsx,json,html}\"",
-    "dev": "cross-env NODE_ENV=development webpack-dev-server",
     "lint": "tslint --project tsconfig.json",
+    "format": "prettier --write \"src/**/*.{ts,tsx,json,html}\"",
+    "build": "rimraf dist && cross-env NODE_ENV=production webpack",
+    "dev": "webpack-dev-server",
     "analyze": "cross-env analyze=true npm run build"
   },
   "author": "zhangist",
@@ -15,19 +15,20 @@
   "license": "MIT",
   "dependencies": {
     "axios": "0.19.0",
-    "i18next": "15.0.7",
-    "i18next-browser-languagedetector": "3.0.1",
-    "react": "16.8.5",
-    "react-dom": "16.8.5",
-    "react-i18next": "10.5.3",
-    "react-redux": "6.0.1",
-    "react-router": "5.0.0",
-    "react-router-dom": "5.0.0",
-    "redux": "4.0.1",
+    "cross-env": "5.2.0",
+    "i18next": "17.0.11",
+    "i18next-browser-languagedetector": "3.0.3",
+    "i18next-xhr-backend": "3.1.2",
+    "react": "16.9.0",
+    "react-dom": "16.9.0",
+    "react-i18next": "10.12.2",
+    "react-redux": "7.1.0",
+    "react-router": "5.0.1",
+    "react-router-dom": "5.0.1",
+    "redux": "4.0.4",
     "redux-thunk": "2.3.0"
   },
   "devDependencies": {
-    "@types/i18next-browser-languagedetector": "3.0.0",
     "@types/node": "12.0.2",
     "@types/react": "16.8.8",
     "@types/react-dom": "16.8.3",
@@ -41,7 +42,6 @@
     "babel-preset-react": "6.24.1",
     "classnames": "2.2.6",
     "copy-webpack-plugin": "5.0.2",
-    "cross-env": "5.2.0",
     "css-loader": "2.1.1",
     "cssnano": "4.1.10",
     "html-webpack-plugin": "3.2.0",
@@ -57,7 +57,7 @@
     "tslint": "5.14.0",
     "tslint-config-prettier": "1.18.0",
     "tslint-react": "3.6.0",
-    "typescript": "3.3.3",
+    "typescript": "3.5.3",
     "webpack": "4.12.2",
     "webpack-bundle-analyzer": "3.3.2",
     "webpack-cli": "3.3.0",
