@@ -1,33 +1,39 @@
-import * as React from "react";
+import React from "react";
 
 export interface State {
   count: number;
 }
 
-export default class PageB extends React.Component<{}, State> {
-  public constructor(props: {}) {
-    super(props);
-    this.state = { count: 0 };
+const PageB: React.FC = () => {
+  const [state, setState] = React.useState<State>({
+    count: 0,
+  });
+
+  if (state.count >= 5) {
+    throw new Error("I am error");
   }
 
-  public render() {
-    if (this.state.count >= 5) {
-      throw new Error("I am error");
-    }
-    return (
+  return (
+    <div>
+      <div>Page B</div>
       <div>
-        <div>Page B</div>
-        <div>
-          <p>Click on the numbers to increase the counters.</p>
-          <p>
-            The counter is programmed to throw when it reaches 5. This simulates
-            a JavaScript error in a component.
-          </p>
-          <span onClick={() => this.setState({ count: this.state.count + 1 })}>
-            {this.state.count}
-          </span>
-        </div>
+        <p>Click on the numbers to increase the counters.</p>
+        <p>
+          The counter is programmed to throw when it reaches 5. This simulates a
+          JavaScript error in a component.
+        </p>
+        <span
+          onClick={() =>
+            setState(prevState => ({
+              ...prevState,
+              count: prevState.count + 1,
+            }))
+          }
+        >
+          {state.count}
+        </span>
       </div>
-    );
-  }
-}
+    </div>
+  );
+};
+export default PageB;
