@@ -1,17 +1,15 @@
-import * as React from "react";
+import React from "react";
 import { Router, Route, Switch } from "react-router-dom";
 import { I18nextProvider } from "react-i18next";
 import { Provider } from "react-redux";
-import { I18nNamespace } from "./enums/i18nNamespace";
 import { ReducerKey } from "./enums/reducerKey";
-import { HistoryService } from "./services/historyService";
-import { I18nService } from "./services/i18nService";
-import { StoreService } from "./services/storeService";
+import { history } from "./common/history";
+import { i18n } from "./common/i18n";
+import { store, injectReducer } from "./common/store";
 import { reducer } from "./store/reducer";
 import ErrorBoundary from "./components/ErrorBoundary";
 import Header from "./components/Header";
 import Footer from "./components/Footer";
-import Loading from "./components/Loading";
 import NotFound from "./components/NotFound";
 import "./components/styles/global.scss";
 import DefaultLoader from "./routes/default/Loader";
@@ -23,67 +21,43 @@ import ThemeDemoLoader from "./routes/theme-demo/Loader";
 import I18nDemoLoader from "./routes/i18n-demo/Loader";
 import UserLoader from "./routes/user/Loader";
 
-const history = HistoryService.getHistory();
-const i18n = I18nService.getI18n();
-const store = StoreService.getStore();
+injectReducer(ReducerKey.App, reducer);
 
-StoreService.injectReducer(ReducerKey.App, reducer);
-
-export interface PageProps {}
-export interface PageState {}
-
-class Page extends React.Component<PageProps, PageState> {
-  public async componentDidMount() {
-    if (!I18nService.hasResourceBundle(I18nNamespace.App)) {
-      try {
-        await I18nService.addResourceBundle(I18nNamespace.App);
-      } catch (error) {
-      } finally {
-        this.forceUpdate();
-      }
-    }
-  }
-
-  public render() {
-    if (!I18nService.hasResourceBundle(I18nNamespace.App)) {
-      return <Loading />;
-    }
-
-    return (
-      <ErrorBoundary>
-        <I18nextProvider i18n={i18n}>
-          <Provider store={store}>
-            <Router history={history}>
-              <div className="app">
-                <div className="header">
-                  <Header />
-                </div>
-                <div className="content">
-                  <Switch>
-                    <Route path="/" exact={true} component={DefaultLoader} />
-                    <Route path="/redux-demo" component={ReduxDemoLoader} />
-                    <Route path="/http-demo" component={HttpDemoLoader} />
-                    <Route path="/pages-demo" component={PagesDemoLoader} />
-                    <Route
-                      path="/url-params-demo"
-                      component={UrlParamsDemoLoader}
-                    />
-                    <Route path="/i18n-demo" component={I18nDemoLoader} />
-                    <Route path="/theme-demo" component={ThemeDemoLoader} />
-                    <Route path="/user" component={UserLoader} />
-                    <Route component={NotFound} />
-                  </Switch>
-                </div>
-                <div className="footer">
-                  <Footer />
-                </div>
+const Page: React.FC = () => {
+  return (
+    <ErrorBoundary>
+      <I18nextProvider i18n={i18n}>
+        <Provider store={store}>
+          <Router history={history}>
+            <div className="app">
+              <div className="header">
+                <Header />
               </div>
-            </Router>
-          </Provider>
-        </I18nextProvider>
-      </ErrorBoundary>
-    );
-  }
-}
+              <div className="content">
+                <Switch>
+                  <Route path="/" exact={true} component={DefaultLoader} />
+                  <Route path="/redux-demo" component={ReduxDemoLoader} />
+                  <Route path="/http-demo" component={HttpDemoLoader} />
+                  <Route path="/pages-demo" component={PagesDemoLoader} />
+                  <Route
+                    path="/url-params-demo"
+                    component={UrlParamsDemoLoader}
+                  />
+                  <Route path="/i18n-demo" component={I18nDemoLoader} />
+                  <Route path="/theme-demo" component={ThemeDemoLoader} />
+                  <Route path="/user" component={UserLoader} />
+                  <Route component={NotFound} />
+                </Switch>
+              </div>
+              <div className="footer">
+                <Footer />
+              </div>
+            </div>
+          </Router>
+        </Provider>
+      </I18nextProvider>
+    </ErrorBoundary>
+  );
+};
 
 export default Page;
