@@ -1,3 +1,3 @@
-import * as React from "react";
+import React from "react";
 
 export default () => <div style={{ padding: "10px" }}>404 Not found!</div>;
