@@ -1,21 +1,5 @@
-export interface RequestConfig {
-  url: string;
-  method?:
-    | "get"
-    | "GET"
-    | "delete"
-    | "DELETE"
-    | "head"
-    | "HEAD"
-    | "options"
-    | "OPTIONS"
-    | "post"
-    | "POST"
-    | "put"
-    | "PUT"
-    | "patch"
-    | "PATCH";
-  headers?: any;
-  params?: any;
-  data?: any;
+import { AxiosRequestConfig } from "axios";
+
+export interface RequestConfig extends AxiosRequestConfig {
+  cancelId?: symbol;
 }
