@@ -1,10 +1,9 @@
 import { ActionType } from "../../../enums/actionType";
-import { Profile } from "../interfaces/profile";
 import { Actions } from "./actions";
 import { counter } from "./states/counter";
 import { friends } from "./states/friends";
 import { hello } from "./states/hello";
-import { profile } from "./states/profile";
+import { profile, Profile } from "./states/profile";
 
 export interface State {
   counter: number;
@@ -12,17 +11,18 @@ export interface State {
   friends: string[];
   profile: Profile;
 }
-export const initialState: State = {
+
+export const getInitialState: () => State = () => ({
   counter,
   friends,
   hello,
   profile,
-};
+});
 
-export function reducer(state = initialState, action: Actions): State {
+export function reducer(state = getInitialState(), action: Actions): State {
   switch (action.type) {
     case ActionType.ReduxDemo_ResetState: {
-      return action.payload || initialState;
+      return action.payload || getInitialState();
     }
     case ActionType.ReduxDemo_UpdateState: {
       return Object.assign({}, state, { ...action.payload });
