@@ -1,4 +1,4 @@
-import * as React from "react";
+import React from "react";
 import { NavLink, Route, Switch } from "react-router-dom";
 import NotFound from "../../components/NotFound";
 import subMenuStyles from "../../components/styles/subMenu.scss";
@@ -6,52 +6,44 @@ import DefaultLoader from "./routes/default/Loader";
 import ProfileLoader from "./routes/profile/Loader";
 import SettingsLoader from "./routes/settings/Loader";
 
-export default class Page extends React.Component {
-  public render() {
-    return (
-      <div>
-        <div style={{ padding: "10px" }}>
-          <NavLink
-            to="/user"
-            exact={true}
-            activeClassName={subMenuStyles.active}
-          >
-            Default
-          </NavLink>
-          <span> / </span>
-          <NavLink
-            to="/user/profile"
-            exact={true}
-            activeClassName={subMenuStyles.active}
-          >
-            Profile
-          </NavLink>
-          <span> / </span>
-          <NavLink
-            to="/user/settings"
-            exact={true}
-            activeClassName={subMenuStyles.active}
-          >
-            Settings
-          </NavLink>
-        </div>
-        <div>
-          <Switch>
-            <Route path="/user" exact={true} component={DefaultLoader} />
-            <Route
-              path="/user/profile"
-              exact={true}
-              component={ProfileLoader}
-            />
-            <Route
-              path="/user/settings"
-              exact={true}
-              component={SettingsLoader}
-            />
-            <Route component={NotFound} />
-          </Switch>
-        </div>
+const Page: React.FC = () => {
+  return (
+    <div>
+      <div style={{ padding: "10px" }}>
+        <NavLink to="/user" exact={true} activeClassName={subMenuStyles.active}>
+          Default
+        </NavLink>
+        <span> / </span>
+        <NavLink
+          to="/user/profile"
+          exact={true}
+          activeClassName={subMenuStyles.active}
+        >
+          Profile
+        </NavLink>
+        <span> / </span>
+        <NavLink
+          to="/user/settings"
+          exact={true}
+          activeClassName={subMenuStyles.active}
+        >
+          Settings
+        </NavLink>
       </div>
-    );
-  }
-}
+      <div>
+        <Switch>
+          <Route path="/user" exact={true} component={DefaultLoader} />
+          <Route path="/user/profile" exact={true} component={ProfileLoader} />
+          <Route
+            path="/user/settings"
+            exact={true}
+            component={SettingsLoader}
+          />
+          <Route component={NotFound} />
+        </Switch>
+      </div>
+    </div>
+  );
+};
+
+export default Page;
