@@ -1,12 +1,12 @@
-import * as React from "react";
+import React from "react";
 
-export default class Page extends React.Component {
-  public render() {
-    return (
-      <div>
-        <div style={{ padding: "10px" }}>User Profile Title</div>
-        <div style={{ padding: "10px" }}>User profile content.</div>
-      </div>
-    );
-  }
-}
+const Page: React.FC = () => {
+  return (
+    <div>
+      <div style={{ padding: "10px" }}>User Profile Title</div>
+      <div style={{ padding: "10px" }}>User profile content.</div>
+    </div>
+  );
+};
+
+export default Page;
