@@ -1,16 +1,16 @@
-import * as React from "react";
+import React from "react";
 import Loading from "../../../../components/Loading";
 
 const Component = React.lazy(() =>
   import(/* webpackChunkName: "route_user_settings" */ "./Page"),
 );
 
-export default class Loader extends React.Component {
-  public render() {
-    return (
-      <React.Suspense fallback={<Loading />}>
-        <Component />
-      </React.Suspense>
-    );
-  }
-}
+const Loader: React.FC = () => {
+  return (
+    <React.Suspense fallback={<Loading />}>
+      <Component />
+    </React.Suspense>
+  );
+};
+
+export default Loader;
